用户权限管理

+++ src/edu/ynu/travel/controller/user/UserController.java
@@ -23,7 +23,7 @@ public class UserController {
     @RequestMapping(value = "/users",method = RequestMethod.POST)
     public SimpleResponse addUser(@RequestBody RegistMessage regist){
         String passwordHash = MD5Util.GetMD5Code(regist.getPassword());
-        short status = (short)0;
+        short status = 0;
         UserEntity userInsert = new UserEntity();
         userInsert.setUsername(regist.getUsername());
         userInsert.setEmail(regist.getEmail());
